Update to ipython 3

+++ nbx/nbmanager/tests/test_gist.py
@@ -4,7 +4,7 @@ from contextlib import contextmanager
 
 import nose.tools as nt
 from IPython.nbformat import current
-from IPython.html.services.notebooks.filenbmanager import FileNotebookManager
+from IPython.html.services.contents.filemanager import FileContentsManager
 from IPython.utils.tempdir import TemporaryDirectory
 
 from ..gist import model_to_files, GistService, Gister
@@ -23,9 +23,10 @@ class TestGist(unittest.TestCase):
 
     def test_model_to_files(self):
         with TemporaryDirectory() as td:
-            fm = FileNotebookManager(notebook_dir=td)
-            model = fm.create_notebook()
-            model = fm.get_notebook(model['name'], model['path'])
+            fm = FileContentsManager(root_dir=td)
+            model = fm.create_file()
+            # need content
+            model = fm.get_model(model['name'], model['path'])
             files = model_to_files(model)
             name = model['name']
             # files should only contain one file
@@ -44,11 +45,13 @@ def create_gist_context(*args, **kwargs):
     """
     contextmanager that creates a gist and makes sure to delete it
     """
+    delete_after = kwargs.pop('delete_after', True)
     gs = GistService()
     gs.login(login, password)
     gist = gs.create_gist(*args, **kwargs)
     yield gist
-    gist.delete()
+    if delete_after:
+        gist.delete()
 
 class TestGistService(unittest.TestCase):
 
@@ -106,6 +109,7 @@ class TestGistService(unittest.TestCase):
 
             # add file
             gist.edit('new desc', files={'new.txt':'new stuff'})
+            # TODO: Need to deepdive and see why the assert fails. #15
             nt.assert_items_equal(gist.files, ['empty.txt', 'new.txt'])
             nt.assert_equals(len(gist.gist.history), 3)
